# docs/api/openapi-mvp3-draft.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ontology Platform Backend",
    "description": "Ontology Platform MVP backend API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/audit-logs": {
      "get": {
        "tags": [
          "Audit"
        ],
        "summary": "List project audit logs",
        "operationId": "list_project_audit_logs_api_v1_projects__project_id__audit_logs_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AuditEventType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          },
          {
            "name": "candidate_kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CandidateKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Candidate Kind"
            }
          },
          {
            "name": "candidate_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Candidate Id"
            }
          },
          {
            "name": "review_task_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Review Task Id"
            }
          },
          {
            "name": "publish_job_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Publish Job Id"
            }
          },
          {
            "name": "actor_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLog"
                  },
                  "title": "Response List Project Audit Logs Api V1 Projects  Project Id  Audit Logs Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/candidates/{candidate_kind}/{candidate_id}/audit-logs": {
      "get": {
        "tags": [
          "Audit"
        ],
        "summary": "List candidate audit logs",
        "operationId": "list_candidate_audit_logs_api_v1_candidates__candidate_kind___candidate_id__audit_logs_get",
        "parameters": [
          {
            "name": "candidate_kind",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CandidateKind"
            }
          },
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLog"
                  },
                  "title": "Response List Candidate Audit Logs Api V1 Candidates  Candidate Kind   Candidate Id  Audit Logs Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError"
      },
      "ApiErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse"
      },
      "AuditEventType": {
        "type": "string",
        "enum": [
          "VALIDATION_JOB_CREATED",
          "VALIDATION_RESULT_RECORDED",
          "REVIEW_TASK_CREATED",
          "REVIEW_TASK_ASSIGNED",
          "CORRECTION_SUBMITTED",
          "REVIEW_DECISION_RECORDED",
          "PUBLISH_JOB_CREATED",
          "PUBLISH_JOB_COMPLETED",
          "PUBLISHED_GRAPH_VERSION_CREATED",
          "PUBLISHED_GRAPH_CURRENT_POINTER_UPDATED"
        ],
        "title": "AuditEventType"
      },
      "AuditLog": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "event_type": {
            "$ref": "#/components/schemas/AuditEventType"
          },
          "actor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Id"
          },
          "candidate_kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidateKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "candidate_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Id"
          },
          "review_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Task Id"
          },
          "review_decision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Decision Id"
          },
          "validation_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Job Id"
          },
          "publish_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publish Job Id"
          },
          "published_graph_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published Graph Version Id"
          },
          "original_snapshot": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Snapshot"
          },
          "corrected_snapshot": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrected Snapshot"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "metadata": {
            "type": "object",
            "title": "Metadata"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "event_type",
          "created_at"
        ],
        "title": "AuditLog"
      },
...
      "CandidateKind": {
        "type": "string",
        "enum": [
          "ENTITY",
          "RELATION",
          "PROPERTY_VALUE"
        ],
        "title": "CandidateKind"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}